Change models module to be a subpackage
Step 1 is just breaking it up as is, but into individual modules.. bc having it all in one file is gonna get nasty especially since all the entities still need to be converted

Step 2 is gonna be an attempt to improve literally everything i guess
