+++ magento/models/__init__.py
@@ -1,0 +1,3 @@
+from .model import Model
+from .product import Product, MediaEntry
+from .category import Category
