+++ magento/models/product.py
@@ -1,142 +1,8 @@
-# Everything without entity_id
 from __future__ import annotations
-import requests
-
-from abc import ABC, abstractmethod
+from magento import clients
 from typing import Union
-from . import clients
-
-
-class Model(ABC):
-    """The base class for all API response wrapper classes
-
-    **Overview**
-
-    * A :class:`Model` is the object representation of any Magento API response
-    * Initialized using the JSON response ``data`` from any API ``endpoint``
-    * Most predefined subclasses use an ``endpoint`` that can be searched with criteria
-    * Access the corresponding :class:`~magento.search.SearchQuery` object via :meth:`~.query_endpoint`
-    """
-
-    def __init__(self, data: dict, client: clients.Client, endpoint: str, private_keys: bool = True):
-        """Initialize a :class:`Model` object from an API response and the ``endpoint`` that it came from
-
-        ...
-
-        **NOTE** that the ``endpoint`` is used to:
-
-        * Generate the :meth:`~.url_for` any requests made by subclass-specific methods
-        * Match the :class:`Model` to its corresponding :class:`~magento.search.SearchQuery` object,
-          which is returned by :meth:`~.query_endpoint`
-        * Determine how to :meth:`~Model.parse` new :class:`Model` objects from API responses
-
-        ...
-
-        :param data: the JSON from an API response to use as source data
-        :param client: an initialized :class:`~.Client`
-        :param endpoint: the base API endpoint that the :class:`Model` represents
-        :param private_keys: if set to True, will set the keys in the :attr:`~.excluded_keys` as private attributes
-            (prefixed with ``__``) instead of fully excluding them
-
-        """
-        if not isinstance(data, dict):
-            raise TypeError(f'Parameter "data" must be of type {dict}')
-        if not isinstance(endpoint, str):
-            raise TypeError(f'Parameter "endpoint" must be of type {str}')
-        if not isinstance(client, clients.Client):
-            raise TypeError(f'Parameter "client" must be of type {clients.Client}')
-
-        self.data = data
-        self.client = client
-        self.endpoint = endpoint
-        self.set_attrs(data, private_keys=private_keys)
-
-    def set_attrs(self, data: dict, private_keys: bool = True) -> None:
-        """Initializes object attributes using the JSON from an API response as the data source
-
-        Called at the time of object initialization, but can also be used to update the source data and
-        reinitialize the attributes without creating a new object
-
-        :param data: the API response JSON to use as the object source data
-        :param private_keys: if set to True, will set the :attr:`~.excluded_keys` as private attributes
-            (prefixed with ``__``) instead of fully excluding them
-
-        **Private Keys Clarification**
-
-        Let's say that ``"status"`` is in the :attr:`~.excluded_keys`
-
-        * No matter what, the :class:`Model` object will not have a ``status`` attribute set
-        * If ``private_keys=True``, it **will** have a ``__status`` attribute set though
-        * If ``private_keys=False``, then the attribute/key is completely excluded
-        """
-        keys = set(data) - set(self.excluded_keys)
-        for key in keys:
-            if key == 'custom_attributes':
-                if attrs := data[key]:
-                    setattr(self, key, self.unpack_attributes(attrs))
-            else:
-                setattr(self, key, data[key])
-
-        if private_keys:
-            private = '_' + self.__class__.__name__ + '__'
-            for key in self.excluded_keys:
-                setattr(self, private + key, data.get(key))
-
-        self.data = data
-
-    @property
-    @abstractmethod
-    def excluded_keys(self) -> list[str]:
-        """API response keys that shouldn't be set as object attributes by :meth:`~.set_attrs`
-
-        :returns: list of API response keys that shouldn't be set as attributes
-        """
-        pass
-
-    def query_endpoint(self):
-        """Initializes and returns the :class:`~.SearchQuery` object corresponding to the Model's ``endpoint``
-
-        :returns: a :class:`~.SearchQuery` or subclass, depending on the ``endpoint``
-        :rtype: :class:`~.SearchQuery`
-        """
-        return self.client.search(self.endpoint)
-
-    def parse(self, response: dict) -> Model:
-        """Uses the instance's corresponding :class:`~.SearchQuery` to parse an API response
-
-        :param response: JSON dictionary from the API to use as source data
-        :return: a :class:`~.Model` object, initialized using the ``response`` as the source data
-        :rtype: :class:`~.Model` with the same ``endpoint`` as the calling instance
-        """
-        return self.query_endpoint().parse(response)
-
-    @staticmethod
-    def unpack_attributes(attributes: list[dict]) -> dict:
-        """Unpacks a list of custom attribute dictionaries into a single dictionary
-
-        **Example**
-
-        >>> custom_attrs = [{'attribute_code': 'attr', 'value': 'val'},{'attribute_code': 'will_to_live', 'value': '0'}]
-        >>> print(Model.unpack_attributes(custom_attrs))
-        {'attr': 'val', 'will_to_live': '0'}
-
-        :param attributes: a list of custom attribute dictionaries
-        :returns: a single dictionary of all custom attributes formatted as ``{"attr": "val"}``
-        """
-        return {attr['attribute_code']: attr['value'] for attr in attributes}
-
-    @staticmethod
-    def encode(string: str) -> str:
-        """URL-encode with :mod:`urllib`; used for requests that could contain special characters
-
-        |  **Example:** requests to the ``products`` endpoint contain a ``sku`` path parameter
-        |      **‣** Since a ``sku`` can contain characters like ``/`` and ``*``, it will always be encoded first
-        |      **‣** See :meth:`~.by_sku` or :attr:`~.encoded_sku`
-
-        :param string: the string to URL-encode
-        """
-        import urllib.parse
-        return urllib.parse.quote_plus(string)
+from . import Model
+import requests
 
 
 class Product(Model):
@@ -223,9 +89,9 @@ class Product(Model):
             if hasattr(entry, attribute):
                 attr = getattr(entry, attribute)
                 if condition == 'in':
-                    criteria = f'{value} {condition} {attr}'    # Ex. "thumbnail" in types
+                    criteria = f'{value} {condition} {attr}'  # Ex. "thumbnail" in types
                 else:
-                    criteria = f'{attr} {condition} {value}'    # Ex. "position" < 2; "id" == 34531
+                    criteria = f'{attr} {condition} {value}'  # Ex. "position" < 2; "id" == 34531
                 try:
                     if eval(criteria):
                         result.append(entry)
@@ -331,7 +197,7 @@ class Product(Model):
     def update_status(self, status):
         if status not in [Product.STATUS_ENABLED, Product.STATUS_DISABLED]:
             raise ValueError('Invalid status provided')
-        
+
         endpoint = f'products/{self.encoded_sku}'
         scopes = (
             self.client.url_for(endpoint),  # No scope also updates default scope
@@ -393,7 +259,6 @@ class Product(Model):
             return False
 
 
-
 class MediaEntry(Model):
     MEDIA_MAPPING = {
         'id': int,
@@ -405,7 +270,7 @@ class MediaEntry(Model):
         'file': str
     }
 
-    def __init__(self, product: Product, entry: dict):
+    def __init__(self, product: "Product", entry: dict):
         super().__init__(
             data=entry,
             client=product.client,
@@ -424,66 +289,3 @@ class MediaEntry(Model):
     @property
     def link(self):
         return f'https://{self.client.domain}/media/catalog/product{self.file}'
-
-
-class Category(Model):
-    DOCUMENTATION = 'https://magento.redoc.ly/2.3.7-admin/tag/categories'
-
-    def __init__(self, data, client):
-        super().__init__(
-            data=data,
-            client=client,
-            endpoint='categories'
-        )
-        self._products = None  # Attributes that are stored after retrieving the first time
-        self._subcategories = None
-
-    def __str__(self):
-        return f'Magento Category: {self.name}'
-
-    @property
-    def excluded_keys(self):
-        return []
-
-    @property
-    def subcategories(self):
-        """Retrieve and temporarily store the category's child categories"""
-        if self._subcategories is None:
-            if hasattr(self, 'children_data'):
-                # children_data: a list of mostly complete subcategory API response dicts
-                self._subcategories = [self.parse(child) for child in self.children_data]
-            elif hasattr(self, 'children'):
-                # children: a comma separated string of subcategory ids; self.subcategory_ids returns them as list
-                self._subcategories = [self.query_endpoint().by_id(child_id) for child_id in self.subcategory_ids]
-            else:
-                self._subcategories = []    # Data comes from those two fields... no subcategories
-        return self._subcategories
-
-    @property
-    def subcategory_ids(self):
-        if hasattr(self, 'children'):
-            return self.children.split(',')
-        else:
-            return [category.id for category in self.subcategories]
-
-    @property
-    def subcategory_names(self):
-        return [category.name for category in self.subcategories]
-
-    @property
-    def products(self):
-        if self._products is None:
-            endpoint = self.client.url_for(f'categories/{self.id}/products')
-            response = self.client.request(endpoint)
-            self._products = [product['sku'] for product in response.json()]
-        return self._products
-
-    @property
-    def all_products(self):
-        """Products of the category AND its subcategories"""
-        products = self.products.copy()
-        for child in self.subcategories:
-            if child.name != 'Default Category':
-                products.extend(child.products)
-        # In case there are products in more than one subcategory
-        return set(products)
